+++ src/types/beatmap/v4/difficulty.ts
@@ -3,6 +3,7 @@ import type { IBombNote } from './bombNote.ts';
 import type { IChain } from './chain.ts';
 import type { IColorNote } from './colorNote.ts';
 import type { IItem } from './item.ts';
+import type { INJSEvent } from './njsEvent.ts';
 import type { IObject, IObjectArc, IObjectChain, IObjectLane } from './object.ts';
 import type { IObstacle } from './obstacle.ts';
 import type { ISpawnRotation } from './spawnRotation.ts';
@@ -11,7 +12,7 @@ import type { ISpawnRotation } from './spawnRotation.ts';
  * Schema for v4 `Difficulty`.
  */
 export interface IDifficulty extends IItem {
-   version: '4.0.0';
+   version: '4.0.0' | '4.1.0';
    colorNotes: IObjectLane[];
    bombNotes: IObjectLane[];
    obstacles: IObjectLane[];
@@ -30,4 +31,6 @@ export interface IDifficulty extends IItem {
     * @deprecated removed as of 1.39, convert to `r` in object lane
     */
    spawnRotationsData?: ISpawnRotation[];
+   njsEvents: IObject[];
+   njsEventData: INJSEvent[];
 }