+++ src/types/beatmap/wrapper/mod.ts
@@ -41,6 +41,7 @@ export * from './lightRotationEventBoxGroup.ts';
 export * from './lightTranslationEvent.ts';
 export * from './lightTranslationEventBox.ts';
 export * from './lightTranslationEventBoxGroup.ts';
+export * from './njsEvent.ts';
 export * from './obstacle.ts';
 export * from './rotationEvent.ts';
 export * from './waypoint.ts';